+++ src/main/java/joshie/harvest/mining/HFMining.java
@@ -102,38 +102,34 @@ public class HFMining {
 
     public static void init() {
         //Spring, Summer, Autumn values
-        MiningContext all = new MiningContext(0);
         MiningContext copper = new MiningContext(COPPER_FLOOR);
         MiningContext silver = new MiningContext(SILVER_FLOOR);
         MiningContext gold = new MiningContext(GOLD_FLOOR);
-        MiningContext gem = new MiningContext(GEM_FLOOR);
         MiningContext mystril = new MiningContext(MYSTRIL_FLOOR);
-        HFApi.mining.registerOre(all, FLOWERS.getStateFromEnum(WEED), 50D, SPRING, SUMMER, AUTUMN);
-        HFApi.mining.registerOre(all, ORE.getStateFromEnum(Ore.ROCK), 100D, SPRING, SUMMER, AUTUMN);
+        HFApi.mining.registerOre(copper, FLOWERS.getStateFromEnum(WEED), 40D, SPRING, SUMMER, AUTUMN);
+        HFApi.mining.registerOre(copper, ORE.getStateFromEnum(Ore.ROCK), 100D, SPRING, SUMMER, AUTUMN);
         HFApi.mining.registerOre(copper, ORE.getStateFromEnum(Ore.COPPER), 6D, SPRING, SUMMER, AUTUMN);
-        HFApi.mining.registerOre(silver, ORE.getStateFromEnum(Ore.SILVER), 5D, SPRING, SUMMER, AUTUMN);
-        HFApi.mining.registerOre(silver, ORE.getStateFromEnum(Ore.AMETHYST), 4D, SPRING, SUMMER, AUTUMN);
-        HFApi.mining.registerOre(gold, ORE.getStateFromEnum(Ore.TOPAZ), 3D, SPRING, SUMMER, AUTUMN);
-        HFApi.mining.registerOre(gold, ORE.getStateFromEnum(Ore.GOLD), 4D, SPRING, SUMMER, AUTUMN);
-        HFApi.mining.registerOre(gem, ORE.getStateFromEnum(Ore.RUBY), 2D, SPRING, SUMMER, AUTUMN);
-        HFApi.mining.registerOre(gem, ORE.getStateFromEnum(Ore.EMERALD), 1D, SPRING, SUMMER, AUTUMN);
-        HFApi.mining.registerOre(mystril, ORE.getStateFromEnum(Ore.MYSTRIL), 3D, SPRING, SUMMER, AUTUMN);
-        HFApi.mining.registerOre(mystril, ORE.getStateFromEnum(Ore.DIAMOND), 0.5D, SPRING, SUMMER, AUTUMN);
+        HFApi.mining.registerOre(copper, ORE.getStateFromEnum(Ore.AMETHYST), 6D, SPRING, SUMMER, AUTUMN);
+        HFApi.mining.registerOre(silver, ORE.getStateFromEnum(Ore.SILVER), 7D, SPRING, SUMMER, AUTUMN);
+        HFApi.mining.registerOre(silver, ORE.getStateFromEnum(Ore.TOPAZ), 5D, SPRING, SUMMER, AUTUMN);
+        HFApi.mining.registerOre(gold, ORE.getStateFromEnum(Ore.GOLD), 8D, SPRING, SUMMER, AUTUMN);
+        HFApi.mining.registerOre(gold, ORE.getStateFromEnum(Ore.RUBY), 6D, SPRING, SUMMER, AUTUMN);
+        HFApi.mining.registerOre(gold, ORE.getStateFromEnum(Ore.EMERALD), 4D, SPRING, SUMMER, AUTUMN);
+        HFApi.mining.registerOre(mystril, ORE.getStateFromEnum(Ore.MYSTRIL), 5D, SPRING, SUMMER, AUTUMN);
+        HFApi.mining.registerOre(mystril, ORE.getStateFromEnum(Ore.DIAMOND), 3D, SPRING, SUMMER, AUTUMN);
 
         //Winter values
-        HFApi.mining.registerOre(all, ORE.getStateFromEnum(Ore.ROCK), 100D, WINTER);
-        HFApi.mining.registerOre(copper, ORE.getStateFromEnum(Ore.COPPER), 10D, WINTER);
-        HFApi.mining.registerOre(copper, ORE.getStateFromEnum(Ore.GEM), 3D, WINTER);
-        HFApi.mining.registerOre(silver, ORE.getStateFromEnum(Ore.SILVER), 6D, WINTER);
-        HFApi.mining.registerOre(silver, ORE.getStateFromEnum(Ore.AMETHYST), 5D, WINTER);
-        HFApi.mining.registerOre(gold, ORE.getStateFromEnum(Ore.TOPAZ), 4D, WINTER);
+        HFApi.mining.registerOre(copper, ORE.getStateFromEnum(Ore.ROCK), 110D, WINTER);
+        HFApi.mining.registerOre(copper, ORE.getStateFromEnum(Ore.COPPER), 9D, WINTER);
+        HFApi.mining.registerOre(copper, ORE.getStateFromEnum(Ore.GEM), 5D, WINTER);
+        HFApi.mining.registerOre(copper, ORE.getStateFromEnum(Ore.AMETHYST), 6D, WINTER);
+        HFApi.mining.registerOre(silver, ORE.getStateFromEnum(Ore.SILVER), 10D, WINTER);
+        HFApi.mining.registerOre(silver, ORE.getStateFromEnum(Ore.TOPAZ), 6D, WINTER);
         HFApi.mining.registerOre(gold, ORE.getStateFromEnum(Ore.GOLD), 10D, WINTER);
-        HFApi.mining.registerOre(gem, ORE.getStateFromEnum(Ore.RUBY), 5D, WINTER);
-        HFApi.mining.registerOre(gem, ORE.getStateFromEnum(Ore.EMERALD), 2D, WINTER);
+        HFApi.mining.registerOre(gold, ORE.getStateFromEnum(Ore.RUBY), 7D, WINTER);
+        HFApi.mining.registerOre(gold, ORE.getStateFromEnum(Ore.EMERALD), 4D, WINTER);
         HFApi.mining.registerOre(mystril, ORE.getStateFromEnum(Ore.MYSTRIL), 5D, WINTER);
-        HFApi.mining.registerOre(mystril, ORE.getStateFromEnum(Ore.DIAMOND), 1D, WINTER);
-
-        //Deeper and deeper values
+        HFApi.mining.registerOre(mystril, ORE.getStateFromEnum(Ore.DIAMOND), 3D, WINTER);
     }
 
     private static void registerSellable(Item item, long value) {