+++ src/main/java/joshie/harvest/shops/HFShops.java
@@ -326,8 +326,8 @@ public class HFShops {
         MINER.addPurchasable(1000, HFMining.LADDER.getStackFromEnum(Ladder.DECORATIVE), 3);
         MINER.addPurchasable(new PurchasableObtainedMaterial(200, HFMining.MINING_TOOL.getStackFromEnum(MiningTool.ELEVATOR_CABLE, 8), Copper.of(1)));
         MINER.addPurchasable(new PurchasableObtainedMaterial(1000, HFMining.ELEVATOR.getStackFromEnum(Elevator.JUNK), Logs.of(3), Copper.of(2), Adamantite.of(1)).setNote(HFNotes.ELEVATOR));
-        MINER.addPurchasable(new PurchasableObtained(200, HFMining.MATERIALS.getStackFromEnum(Material.COPPER)));
-        MINER.addPurchasable(new PurchasableObtained(300, HFMining.MATERIALS.getStackFromEnum(Material.SILVER)));
+        MINER.addPurchasable(new PurchasableObtained(100, HFMining.MATERIALS.getStackFromEnum(Material.COPPER)));
+        MINER.addPurchasable(new PurchasableObtained(200, HFMining.MATERIALS.getStackFromEnum(Material.SILVER)));
         MINER.addPurchasable(new PurchasableObtained(400, HFMining.MATERIALS.getStackFromEnum(Material.GOLD)));
         MINER.addPurchasable(new PurchasableOre(150, new ItemStack(Items.COAL)));
         MINER.addPurchasable(new PurchasableOre(400, new ItemStack(Items.IRON_INGOT)));